requirements.txt
torchtext==0.2.3
ruamel.yaml
ipdb